Creating a site uses wpcom-request-controls
