+++ packages/data-stores/src/site/index.ts
@@ -8,11 +8,11 @@ import { registerStore } from '@wordpress/data';
  */
 import { STORE_KEY } from './constants';
 import reducer, { State } from './reducer';
-import * as actions from './actions';
+import { createActions } from './actions';
 import * as selectors from './selectors';
-import createControls from './controls';
 import { DispatchFromMap, SelectFromMap } from '../mapped-types';
 import { WpcomClientCredentials } from '../shared-types';
+import { controls } from '../wpcom-request-controls';
 
 export * from './types';
 export { State };
@@ -20,13 +20,11 @@ export { State };
 let isRegistered = false;
 export function register( clientCreds: WpcomClientCredentials ): typeof STORE_KEY {
 	if ( ! isRegistered ) {
-		const controls = createControls( clientCreds );
 		isRegistered = true;
 		registerStore< State >( STORE_KEY, {
-			actions,
+			actions: createActions( clientCreds ),
 			controls: controls as any,
-			reducer: reducer as any,
-			resolvers: {},
+			reducer,
 			selectors,
 		} );
 	}
@@ -34,6 +32,6 @@ export function register( clientCreds: WpcomClientCredentials ): typeof STORE_KE
 }
 
 declare module '@wordpress/data' {
-	function dispatch( key: typeof STORE_KEY ): DispatchFromMap< typeof actions >;
+	function dispatch( key: typeof STORE_KEY ): DispatchFromMap< ReturnType< typeof createActions > >;
 	function select( key: typeof STORE_KEY ): SelectFromMap< typeof selectors >;
 }
