+++ packages/data-stores/src/user/resolvers.ts
@@ -1,8 +1,9 @@
 /**
  * Internal dependencies
  */
-import { wpcomRequest, WpcomClientCredentials } from '../wpcom-request-controls';
+import { wpcomRequest } from '../wpcom-request-controls';
 import { createActions } from './actions';
+import { WpcomClientCredentials } from '../shared-types';
 
 export function createResolvers( clientCreds: WpcomClientCredentials ) {
 	const { receiveCurrentUser, receiveCurrentUserFailed } = createActions( clientCreds );
